requirements.txt
flask==1.0.2
flask-basicauth==0.2.0
pyyaml==5.1
green==2.14.2
flask-mysqldb==0.2.0
